bank: make ListingSpecification#isLatest ~24040.38 times cheaper
according to the pgsql query planner, using an independent query in
order to yield a "last refresh" value for all securities produces a
query with cost 2885.25 when analyzed against our current dataset,
whereas the cost of the old query that this specification would create
is 69362511.78.  do note that the latter lacks the analysis based on
data-set, as I was forced to use EXPLAIN over EXPLAIN ANALYSE because
the EXPLAIN ANALYSE I started has been running for an hour at the time
of writing.

this query is suboptimal, but we're dealing with a limitation of JQL: we
cannot select tuples in subqueries to check against.  this query is, as
a result, approximate, unless lastRefresh are unique, which there is no
guarantee of (but a high likelihood of regardless).  oh well.  I don't
care enough to fix it.